Use detailed flight lookup
Complete rewrite of the function for getting flight details. Use chunking to access the flightradar flight ID info that is so large it otherwise crashes the board. Expose a whole bunch of fields about the flight and make it easier to change which ones are used. Remove the icao file lookup since aircraft model is now part of the same code.